common cli-tests utils
