+++ cli-tests/src/main.rs
@@ -1,6 +1,8 @@
 #[cfg(test)]
 mod upload;
 #[cfg(test)]
+mod utils;
+#[cfg(test)]
 mod validate;
 
 fn main() {}
