+++ cli-tests/src/upload.rs
@@ -1,9 +1,8 @@
-use std::{
-    env, fs,
-    io::BufReader,
-    path::{Path, PathBuf},
-};
+use std::{fs, io::BufReader};
 
+use crate::utils::{
+    generate_mock_codeowners, generate_mock_git_repo, generate_mock_valid_junit_xmls, CARGO_RUN,
+};
 use api::{
     BundleUploadStatus, CreateRepoRequest, GetQuarantineBulkTestStatusRequest,
     UpdateBundleUploadRequest,
@@ -11,56 +10,19 @@ use api::{
 use assert_cmd::Command;
 use assert_matches::assert_matches;
 use context::repo::RepoUrlParts as Repo;
-use escargot::{CargoBuild, CargoRun};
-use junit_mock::JunitMock;
-use lazy_static::lazy_static;
 use tempfile::tempdir;
-use test_utils::{
-    mock_git_repo::setup_repo_with_commit,
-    mock_server::{spawn_mock_server, RequestPayload},
-};
+use test_utils::mock_server::{spawn_mock_server, RequestPayload};
 use trunk_analytics_cli::{
     codeowners::CodeOwners,
     types::{BundleMeta, FileSetType},
 };
-
-lazy_static! {
-    static ref CARGO_MANIFEST_DIR: PathBuf = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    pub static ref CARGO_RUN: CargoRun = CargoBuild::new()
-        .bin("trunk-analytics-cli")
-        .target_dir(CARGO_MANIFEST_DIR.join("../target"))
-        .manifest_path(CARGO_MANIFEST_DIR.join("../cli/Cargo.toml"))
-        .features("force-sentry-env-dev")
-        .current_release()
-        .current_target()
-        .run()
-        .unwrap();
-}
-
-fn generate_mock_git_repo<T: AsRef<Path>>(directory: T) {
-    setup_repo_with_commit(directory).unwrap();
-}
-
-fn generate_mock_junit_xmls<T: AsRef<Path>>(directory: T) {
-    let mut jm = JunitMock::new(junit_mock::Options::default());
-    let reports = jm.generate_reports();
-    JunitMock::write_reports_to_file(directory.as_ref(), reports).unwrap();
-}
-
-fn generate_mock_codeowners<T: AsRef<Path>>(directory: T) {
-    const CODEOWNERS: &str = r#"
-        [Owners of Everything]
-        * @user
-    "#;
-    fs::write(directory.as_ref().join("CODEOWNERS"), CODEOWNERS).unwrap();
-}
 
 // NOTE: must be multi threaded to start a mock server
 #[tokio::test(flavor = "multi_thread")]
 async fn upload_bundle() {
     let temp_dir = tempdir().unwrap();
     generate_mock_git_repo(&temp_dir);
-    generate_mock_junit_xmls(&temp_dir);
+    generate_mock_valid_junit_xmls(&temp_dir);
     generate_mock_codeowners(&temp_dir);
 
     let state = spawn_mock_server().await;
